+++ pom.xml
@@ -329,9 +329,10 @@ LICENSE file in the root directory of this source tree.
 								<version>${jline.version}</version>
 							</dependency>
 							<dependency>
-								<groupId>org.jline</groupId>
-								<artifactId>jline-terminal-jansi</artifactId>
-								<version>${jline.version}</version>
+								<groupId>org.fusesource.jansi</groupId>
+								<artifactId>jansi</artifactId>
+								<version>${jansi.version}</version>
+								<optional>true</optional>
 							</dependency>
 						</dependencies>
 					</plugin>
@@ -365,9 +366,9 @@ LICENSE file in the root directory of this source tree.
 			<optional>true</optional>
 		</dependency>
 		<dependency>
-			<groupId>org.jline</groupId>
-			<artifactId>jline-terminal-jansi</artifactId>
-			<version>${jline.version}</version>
+			<groupId>org.fusesource.jansi</groupId>
+			<artifactId>jansi</artifactId>
+			<version>1.18</version>
 			<optional>true</optional>
 		</dependency>
 
@@ -427,6 +428,7 @@ LICENSE file in the root directory of this source tree.
 		<commons-lang3.version>3.5</commons-lang3.version>
 		<ognl.version>3.1.23</ognl.version>
 		<jline.version>3.11.0</jline.version>
+		<jansi.version>1.18</jansi.version>
 		<slf4j.version>1.7.26</slf4j.version>
 		<logback.version>1.2.3</logback.version>
 		<junit.version>4.12</junit.version>
