+++ src/main/java/jp/co/future/uroborosql/client/SqlREPL.java
@@ -22,44 +22,33 @@ import java.sql.Connection;
 import java.sql.DatabaseMetaData;
 import java.sql.Driver;
 import java.sql.DriverManager;
-import java.sql.DriverPropertyInfo;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.SQLFeatureNotSupportedException;
-import java.sql.Time;
-import java.sql.Timestamp;
-import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Enumeration;
 import java.util.HashMap;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 import java.util.ServiceLoader;
-import java.util.Set;
-import java.util.SortedSet;
-import java.util.TreeSet;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import org.apache.commons.lang3.BooleanUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.math.NumberUtils;
-import org.apache.commons.lang3.time.DateUtils;
-import org.jline.reader.Candidate;
 import org.jline.reader.Completer;
 import org.jline.reader.EndOfFileException;
 import org.jline.reader.LineReader;
 import org.jline.reader.LineReaderBuilder;
-import org.jline.reader.ParsedLine;
 import org.jline.reader.UserInterruptException;
+import org.jline.reader.impl.DefaultHighlighter;
 import org.jline.reader.impl.completer.AggregateCompleter;
 import org.jline.terminal.Terminal;
 import org.jline.terminal.TerminalBuilder;
+import org.jline.utils.AttributedStringBuilder;
+import org.jline.utils.AttributedStyle;
 import org.jline.utils.InfoCmp.Capability;
 import org.slf4j.LoggerFactory;
 
@@ -67,6 +56,11 @@ import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
 import jp.co.future.uroborosql.SqlAgent;
 import jp.co.future.uroborosql.UroboroSQL;
+import jp.co.future.uroborosql.client.completer.BindParamCompleter;
+import jp.co.future.uroborosql.client.completer.ReplCommandCompleter;
+import jp.co.future.uroborosql.client.completer.SqlKeywordCompleter;
+import jp.co.future.uroborosql.client.completer.SqlNameCompleter;
+import jp.co.future.uroborosql.client.completer.TableNameCompleter;
 import jp.co.future.uroborosql.config.SqlConfig;
 import jp.co.future.uroborosql.context.SqlContext;
 import jp.co.future.uroborosql.context.SqlContextFactory;
@@ -75,18 +69,7 @@ import jp.co.future.uroborosql.filter.DumpResultSqlFilter;
 import jp.co.future.uroborosql.filter.SqlFilterManagerImpl;
 import jp.co.future.uroborosql.mapping.Table;
 import jp.co.future.uroborosql.mapping.TableMetadata;
-import jp.co.future.uroborosql.node.BindVariableNode;
-import jp.co.future.uroborosql.node.EmbeddedValueNode;
-import jp.co.future.uroborosql.node.IfNode;
-import jp.co.future.uroborosql.node.Node;
-import jp.co.future.uroborosql.node.ParenBindVariableNode;
-import jp.co.future.uroborosql.parser.ContextTransformer;
-import jp.co.future.uroborosql.parser.SqlParser;
-import jp.co.future.uroborosql.parser.SqlParserImpl;
 import jp.co.future.uroborosql.store.NioSqlManagerImpl;
-import ognl.ASTProperty;
-import ognl.Ognl;
-import ognl.OgnlException;
 
 /**
  * SQL REPL実装クラス
@@ -96,129 +79,12 @@ import ognl.OgnlException;
  * @author H.Sugimoto
  */
 public class SqlREPL {
-	/** バインドパラメータ中の定数指定を判定するための正規表現 */
-	private static final Pattern CONSTANT_PAT = Pattern.compile("^[A-Z][A-Z0-9_-]*$");
-
 	/** プロパティ上のクラスパスに指定された環境変数を置換するための正規表現 */
 	private static final Pattern SYSPROP_PAT = Pattern.compile("\\$\\{(.+?)\\}");
 
 	/** DESCで表示する項目 */
 	private static final String[] DESC_COLUMN_LABELS = { "TABLE_NAME", "COLUMN_NAME", "TYPE_NAME", "COLUMN_SIZE",
 			"DECIMAL_DIGITS", "IS_NULLABLE", "COLUMN_DEF", "REMARKS" };
-
-	/** コマンド名 */
-	@SuppressWarnings("unchecked")
-	private enum Command {
-		/** SQLの検索実行 */
-		QUERY(false, SqlNameCompleter.class, BindParamCompleter.class),
-		/** SQLの更新実行 */
-		UPDATE(false, SqlNameCompleter.class, BindParamCompleter.class),
-		/** SQLファイルの参照 */
-		VIEW(false, SqlNameCompleter.class),
-		/** SQLファイルのリスト出力 */
-		LIST(false, SqlNameCompleter.class),
-		/** 入力コマンドの履歴出力 */
-		HISTORY(false),
-		/** リロード */
-		RELOAD(false),
-		/** 登録されているJDBCドライバーのリスト出力 */
-		DRIVER(false),
-		/** テーブル定義表示 */
-		DESC(false, TableNameCompleter.class),
-		/** SQL文生成 */
-		GENERATE(false, SqlKeywordCompleter.class, TableNameCompleter.class),
-		/** ヘルプメッセージ出力 */
-		HELP(false),
-		/** 画面のクリア */
-		CLS(false),
-		/** SqlREPLの終了 */
-		EXIT(false),
-		/** スペシャルメッセージ */
-		THIS(true);
-
-		/** 利用する入力補完の並び */
-		private List<Class<? extends Completer>> completers = new ArrayList<>();
-		/** HELPコマンドで非表示 */
-		private boolean hidden = false;
-
-		/**
-		 * 入力コマンド
-		 *
-		 * @param hidden HELPコマンドで表示しない場合に<code>true</code>
-		 * @param completers 補完器名
-		 */
-		private Command(final boolean hidden, final Class<? extends Completer>... completers) {
-			this.hidden = hidden;
-			this.completers = Arrays.asList(completers);
-		}
-
-		/**
-		 * 指定されたコード補完の対象とする引数の開始位置を返却する
-		 *
-		 * @param completer 対象とするコード補完名
-		 * @return 引数の位置。該当するコード補完がない場合は<code>-1</code>を返す
-		 */
-		public int getStartArgNo(final Class<? extends Completer> completer) {
-			int idx = completers.indexOf(completer);
-			return idx >= 0 ? idx + 1 : -1;
-		}
-
-		/**
-		 * 非表示コマンドかどうか
-		 * @return 非表示コマンドの場合<code>true</code>
-		 */
-		public boolean isHidden() {
-			return hidden;
-		}
-
-		/**
-		 * 指定された入力がコマンドにマッチするかどうかを判定する
-		 * @param input 入力文字列
-		 * @return コマンドにマッチする場合<code>true</code>
-		 */
-		public boolean match(final String input) {
-			return !isHidden() && toString().startsWith(input.toLowerCase());
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see java.lang.Enum#toString()
-		 */
-		@Override
-		public String toString() {
-			return super.toString().toLowerCase();
-		}
-
-		/**
-		 * 文字列からCommandを取得する
-		 * @param cmd コマンド文字列
-		 * @return 対応するCommandクラス。該当するCommandがない場合、HelpCommandを返却する
-		 */
-		public static Command toCommand(final String cmd) {
-			if (StringUtils.isNotBlank(cmd)) {
-				for (Command command : values()) {
-					if (command.toString().equalsIgnoreCase(cmd.trim())) {
-						return command;
-					}
-				}
-			}
-			return HELP;
-		}
-
-		public static Completer getCommandCompleter() {
-			return new Completer() {
-				@Override
-				public void complete(final LineReader reader, final ParsedLine line, final List<Candidate> candidates) {
-					if (line.wordIndex() == 0) {
-						Stream.of(Command.values()).filter(c -> c.match(line.word()))
-								.map(c -> new Candidate(c.toString()))
-								.forEach(candidates::add);
-					}
-				}
-			};
-		}
-	}
 
 	/** プロパティパス */
 	private final Path propPath;
@@ -230,7 +96,7 @@ public class SqlREPL {
 	private URLClassLoader additionalClassLoader = null;
 
 	/** SQL設定クラス */
-	private SqlConfig config = null;
+	SqlConfig config = null;
 
 	/**
 	 * メインメソッド
@@ -297,7 +163,7 @@ public class SqlREPL {
 			showMessage(terminal, "/message.txt");
 			showProps(terminal);
 
-			initialize();
+			initialize(terminal);
 
 			listen(terminal);
 
@@ -308,9 +174,12 @@ public class SqlREPL {
 	/**
 	 * 初期化処理
 	 *
+	 * @param terminal Terminal
 	 * @throws Exception 実行時例外
 	 */
-	private void initialize() throws Exception {
+	private void initialize(final Terminal terminal) throws Exception {
+		terminal.writer().println("initialize.");
+
 		ClassLoader currentClassLoader = Thread.currentThread().getContextClassLoader();
 
 		if (currentClassLoader.equals(additionalClassLoader)) {
@@ -402,7 +271,7 @@ public class SqlREPL {
 
 	/**
 	 * 終了時処理
-	 * @param terminal ターミナル
+	 * @param terminal Terminal
 	 */
 	private void dispose(final Terminal terminal) {
 		terminal.writer().println("SQL REPL end.");
@@ -419,20 +288,20 @@ public class SqlREPL {
 	/**
 	 * コンソールからの入力のリスン
 	 *
-	 * @param terminal ターミナル
+	 * @param terminal Terminal
 	 * @throws IOException I/O例外
 	 */
 	private void listen(final Terminal terminal) throws IOException {
 		List<Completer> completers = new ArrayList<>();
 
 		// コマンドのコード補完
-		completers.add(Command.getCommandCompleter());
+		completers.add(new ReplCommandCompleter());
 		// SQL名のコード補完
-		completers.add(new SqlNameCompleter(config.getSqlManager().getSqlPathList()));
+		completers.add(new SqlNameCompleter(config.getSqlManager()));
 		// バインドパラメータのコード補完
-		completers.add(new BindParamCompleter());
+		completers.add(new BindParamCompleter(config.getSqlManager()));
 		// テーブル名のコード補完
-		completers.add(new TableNameCompleter());
+		completers.add(new TableNameCompleter(config.getConnectionSupplier()));
 		// SQLキーワードのコード補完
 		completers.add(new SqlKeywordCompleter());
 
@@ -440,6 +309,7 @@ public class SqlREPL {
 				.appName("uroborosql")
 				.terminal(terminal)
 				.completer(new AggregateCompleter(completers))
+				.highlighter(new DefaultHighlighter())
 				.option(LineReader.Option.CASE_INSENSITIVE, true)
 				.build();
 		while (true) { // ユーザの一行入力を待つ
@@ -447,9 +317,7 @@ public class SqlREPL {
 				if (!commandExecute(reader)) {
 					break;
 				}
-			} catch (UserInterruptException ex) {
-				// do nothing
-			} catch (EndOfFileException ex) {
+			} catch (UserInterruptException | EndOfFileException ex) {
 				break;
 			} catch (Exception e) {
 				e.printStackTrace(System.err);
@@ -465,15 +333,20 @@ public class SqlREPL {
 	 * @throws Exception 実行時例外
 	 */
 	private boolean commandExecute(final LineReader reader) throws Exception {
-		String prompt = "uroborosql > ";
+		String prompt = new AttributedStringBuilder()
+				.style(AttributedStyle.BOLD.foreground(AttributedStyle.GREEN))
+				.ansiAppend("uroborosql")
+				.style(AttributedStyle.DEFAULT)
+				.append(" > ")
+				.toAnsi();
 		String line = reader.readLine(prompt);
 		if (line == null || line.isEmpty()) {
 			// 空なら何もせずにループ
 			return true;
 		}
 
-		String[] parts = getLineParts(line);
-		Command command = Command.toCommand(parts[0]);
+		String[] parts = line.split("\\s+");
+		ReplCommand command = ReplCommand.toCommand(parts[0]);
 
 		PrintWriter writer = reader.getTerminal().writer();
 
@@ -493,7 +366,7 @@ public class SqlREPL {
 			}
 			props = loadProps(propPath);
 			showProps(reader.getTerminal());
-			initialize();
+			initialize(reader.getTerminal());
 
 			return true;
 
@@ -642,8 +515,9 @@ public class SqlREPL {
 			writer.flush();
 
 			if (parts.length < 3) {
-				writer.println(Command.GENERATE.toString() + " parameter missing. " + Command.GENERATE.toString()
-						+ " [SQL_KEYWORD] [TABLE NAME].");
+				writer.println(
+						ReplCommand.GENERATE.toString() + " parameter missing. " + ReplCommand.GENERATE.toString()
+								+ " [SQL_KEYWORD] [TABLE NAME].");
 				return true;
 			}
 
@@ -732,7 +606,7 @@ public class SqlREPL {
 						SqlContext ctx = agent.contextFrom(sqlName);
 						ctx.setSql(config.getSqlManager().getSql(ctx.getSqlName()));
 						String[] params = Arrays.copyOfRange(parts, 2, parts.length);
-						setSqlParams(ctx, params);
+						SqlParamUtils.setSqlParams(ctx, params);
 
 						ctx.setResultSetType(ResultSet.TYPE_SCROLL_INSENSITIVE);
 
@@ -760,7 +634,7 @@ public class SqlREPL {
 						SqlContext ctx = agent.contextFrom(sqlName);
 						ctx.setSql(config.getSqlManager().getSql(ctx.getSqlName()));
 						String[] params = Arrays.copyOfRange(parts, 2, parts.length);
-						setSqlParams(ctx, params);
+						SqlParamUtils.setSqlParams(ctx, params);
 						try {
 							int ans = agent.update(ctx);
 							agent.commit();
@@ -780,233 +654,6 @@ public class SqlREPL {
 		default:
 			return true;
 		}
-	}
-
-	/**
-	 * 入力行をパーツに分解する。空白で分解した後に条件によって結合を行う。<br>
-	 *
-	 * 【結合条件】<br>
-	 * ・''で囲まれる値には空白を含めることができる。ex) 'String value includes whitespace.'<br>
-	 * ・[]で囲まれる値は配列を表す。 ex) [10, 20, 30, 40], ['aaa' , 'bbb' , 'ccc']
-	 *
-	 * @param line 入力行
-	 * @return パーツに分解した文字列配列
-	 */
-	private String[] getLineParts(final String line) {
-		String[] parts = line.split(" ");
-		List<String> ans = new ArrayList<>();
-		int idx = 0;
-		int len = parts.length;
-		while (idx < len) {
-			if (parts[idx].contains("='") && !parts[idx].endsWith("'")) {
-				StringBuilder builder = new StringBuilder(parts[idx++]);
-				while (idx < len) {
-					builder.append(" ").append(parts[idx]);
-					if (parts[idx++].endsWith("'")) {
-						break;
-					}
-				}
-				ans.add(builder.toString());
-			} else if (parts[idx].contains("=[") && !parts[idx].endsWith("]")) {
-				StringBuilder builder = new StringBuilder(parts[idx++]);
-				while (idx < len) {
-					builder.append(" ").append(parts[idx]);
-					if (parts[idx++].endsWith("]")) {
-						break;
-					}
-				}
-				ans.add(builder.toString());
-			} else {
-				String part = parts[idx++];
-				if (StringUtils.isNotEmpty(part)) {
-					ans.add(part);
-				}
-			}
-		}
-		return ans.toArray(new String[ans.size()]);
-	}
-
-	/**
-	 * SQLバインドパラメータを設定する
-	 *
-	 * @param ctx SQLコンテキスト
-	 * @param paramsArray パラメータ配列
-	 */
-	private void setSqlParams(final SqlContext ctx, final String[] paramsArray) {
-		Set<String> bindParams = getSqlParams(ctx.getSql());
-
-		for (String element : paramsArray) {
-			String[] param = element.split("=");
-			String key = param[0];
-			if (bindParams.remove(key)) {
-				// キーがバインドパラメータに存在するときは値を設定する
-				if (param.length == 1) {
-					// キーだけの指定は値をnullと扱う
-					ctx.param(key, null);
-				} else {
-					String val = param[1];
-					setParam(ctx, key, val);
-				}
-			}
-		}
-
-		// 指定がなかったキーについてはnullを設定する
-		bindParams.forEach(s -> ctx.param(s, null));
-	}
-
-	/**
-	 * 1つのパラメータの設定
-	 *
-	 * パラメータ値は以下の表記が可能
-	 * <dl>
-	 * 	<dh>[NULL]</dh>
-	 *  <dd><code>null</code>を設定する</dd>
-	 * 	<dh>[EMPTY]</dh>
-	 *  <dd>""（空文字）を設定する</dd>
-	 *  <dh>'値'</dh>
-	 *  <dd>文字列として設定する. 空白を含めることもできる</dd>
-	 * 	<dh>[値1,値2,...]</dh>
-	 *  <dd>配列として設定する</dd>
-	 * 	<dh>その他</dh>
-	 *  <dd>文字列として設定する</dd>
-	 * </dl>
-	 *
-	 * @param ctx SqlContext
-	 * @param key パラメータキー
-	 * @param val パラメータ値
-	 */
-	private void setParam(final SqlContext ctx, final String key, final String val) {
-		if (val.startsWith("[") && val.endsWith("]")) {
-			// [] で囲まれた値は配列に変換する。ex) [1, 2] => {"1", "2"}
-			String[] parts = val.substring(1, val.length() - 1).split("\\s*,\\s*");
-			Object[] vals = new Object[parts.length];
-			for (int i = 0; i < parts.length; i++) {
-				vals[i] = convertSingleValue(parts[i]);
-			}
-			ctx.paramList(key, vals);
-		} else {
-			ctx.param(key, convertSingleValue(val));
-		}
-	}
-
-	/**
-	 * パラメータで渡された単独の値を型変換する
-	 *
-	 * @param val 値の文字列
-	 * @return 変換後オブジェクト
-	 */
-	private Object convertSingleValue(final String val) {
-		String value = StringUtils.trim(val);
-		if (StringUtils.isEmpty(value)) {
-			return null;
-		} else if ("[NULL]".equalsIgnoreCase(value)) {
-			return null;
-		} else if ("[EMPTY]".equalsIgnoreCase(value)) {
-			return "";
-		} else if (value.startsWith("'") && value.endsWith("'")) {
-			// ''で囲まれた値は文字列として扱う。空白を含むこともできる。 ex) 'This is a pen'
-			return value.substring(1, value.length() - 1);
-		} else if (Boolean.TRUE.toString().equalsIgnoreCase(value)) {
-			return Boolean.TRUE;
-		} else if (Boolean.FALSE.toString().equalsIgnoreCase(value)) {
-			return Boolean.FALSE;
-		} else if (NumberUtils.isCreatable(value)) {
-			return NumberUtils.createNumber(value);
-		} else {
-			try {
-				// 日時に変換できるか
-				return new Timestamp(DateUtils.parseDateStrictly(value, "yyyy-MM-dd'T'HH:mm:ss").getTime());
-			} catch (ParseException ex) {
-				// do nothing
-			}
-
-			try {
-				// 日付に変換できるか？
-				return new java.sql.Date(DateUtils.parseDateStrictly(value, "yyyy-MM-dd").getTime());
-			} catch (ParseException ex) {
-				// do nothing
-			}
-
-			try {
-				// 時刻に変換できるか？
-				return new Time(DateUtils.parseDateStrictly(value, "HH:mm:ss").getTime());
-			} catch (ParseException ex) {
-				// do nothing
-			}
-			return value;
-		}
-	}
-
-	/**
-	 * SQLパラメータの解析
-	 *
-	 * @param sql 解析対象SQL
-	 * @return SQLを解析して取得したパラメータキーのセット
-	 */
-	private Set<String> getSqlParams(final String sql) {
-		SqlParser parser = new SqlParserImpl(sql);
-		ContextTransformer transformer = parser.parse();
-		Node rootNode = transformer.getRoot();
-
-		Set<String> params = new LinkedHashSet<>();
-		traverseNode(rootNode, params);
-
-		params.removeIf(s -> CONSTANT_PAT.matcher(s).matches());
-
-		return params;
-	}
-
-	/**
-	 * SQLの探索
-	 *
-	 * @param node SQLノード
-	 * @param params パラメータが見つかった場合に格納するSetオブジェクト
-	 */
-	private void traverseNode(final Node node, final Set<String> params) {
-		if (node instanceof BindVariableNode) {
-			params.add(((BindVariableNode) node).getExpression());
-		} else if (node instanceof ParenBindVariableNode) {
-			params.add(((ParenBindVariableNode) node).getExpression());
-		} else if (node instanceof EmbeddedValueNode) {
-			params.add(((EmbeddedValueNode) node).getExpression());
-		} else if (node instanceof IfNode) {
-			try {
-				String expression = ((IfNode) node).getExpression();
-				ognl.Node ognlNode = (ognl.Node) Ognl.parseExpression(expression);
-				traverseExpression(ognlNode, params);
-			} catch (OgnlException e) {
-				e.printStackTrace();
-			}
-		}
-
-		for (int i = 0; i < node.getChildSize(); i++) {
-			Node childNode = node.getChild(i);
-			traverseNode(childNode, params);
-		}
-
-	}
-
-	/**
-	 * 評価式の探索
-	 *
-	 * @param node SQLノード
-	 * @param params パラメータが見つかった場合に格納するSetオブジェクト
-	 */
-	private void traverseExpression(final ognl.Node node, final Set<String> params) {
-		if (node == null) {
-			return;
-		}
-		if (node instanceof ASTProperty) {
-			ASTProperty prop = (ASTProperty) node;
-			params.add(prop.toString());
-		} else {
-			int childCount = node.jjtGetNumChildren();
-			for (int i = 0; i < childCount; i++) {
-				ognl.Node child = node.jjtGetChild(i);
-				traverseExpression(child, params);
-			}
-		}
-
 	}
 
 	/**
@@ -1043,336 +690,6 @@ public class SqlREPL {
 		});
 		terminal.writer().println();
 		terminal.flush();
-	}
-
-	/**
-	 * SQL Name の補完を行うCompleter
-	 *
-	 * @author H.Sugimoto
-	 *
-	 */
-	class SqlNameCompleter implements Completer {
-		private final SortedSet<String> sqlNames = new TreeSet<>();
-
-		/**
-		 * コンストラクタ
-		 * @param sqlNames SQL名のリスト
-		 */
-		private SqlNameCompleter(final List<String> sqlNames) {
-			this.sqlNames.addAll(sqlNames);
-		}
-
-		/**
-		 * コンストラクタから渡されたSQL名の一覧と入力を比較し、前方一致するものがあれば補完対象とする<br>
-		 *
-		 * {@inheritDoc}
-		 *
-		 * @see org.jline.reader.Completer#complete(org.jline.reader.LineReader, org.jline.reader.ParsedLine, java.util.List)
-		 */
-		@Override
-		public void complete(final LineReader reader, final ParsedLine line, final List<Candidate> candidates) {
-			String buffer = line.line();
-			String[] parts = getLineParts(buffer);
-			int len = parts.length;
-
-			// コード補完する引数の番号を特定。
-			int startArgNo = len >= 1 ? Command.toCommand(parts[0]).getStartArgNo(this.getClass()) : -1;
-
-			// 対象引数が-1の時は該当なしなのでコード補完しない
-			if (startArgNo == -1) {
-				return;
-			} else {
-				boolean isBlank = buffer.endsWith(" ");
-				if (len == startArgNo && isBlank || len == startArgNo + 1 && !isBlank) {
-					// コマンドが引数ありの場合
-					String args = len == startArgNo + 1 ? parts[startArgNo] : "";
-					if (StringUtils.isEmpty(args)) {
-						candidates.addAll(sqlNames.stream().map(Candidate::new).collect(Collectors.toList()));
-					} else {
-						for (String match : sqlNames.tailSet(args)) {
-							if (!match.startsWith(args)) {
-								break;
-							}
-							candidates.add(new Candidate(match));
-						}
-					}
-				}
-			}
-		}
-	}
-
-	/**
-	 * バインドパラメータを補完するCompleter
-	 *
-	 * @author H.Sugimoto
-	 *
-	 */
-	class BindParamCompleter implements Completer {
-		/**
-		 * バッファから渡されたSQL名のSQLを取得し、バインドパラメータをパースして取得する。<br>
-		 * 取得したバインドパラメータと入力を比較し、前方一致する場合は補完候補にする。<br>
-		 * ただし、すでに指定済みのバインドパラメータは補完候補から除く<br>
-		 *
-		 * {@inheritDoc}
-		 *
-		 * @see org.jline.reader.Completer#complete(org.jline.reader.LineReader, org.jline.reader.ParsedLine, java.util.List)
-		 */
-		@Override
-		public void complete(final LineReader reader, final ParsedLine line, final List<Candidate> candidates) {
-			String buffer = line.line();
-			String[] parts = getLineParts(buffer);
-			int pos = buffer.length();
-			int len = parts.length;
-
-			// コード補完する引数の番号を特定。
-			int startArgNo = len >= 1 ? Command.toCommand(parts[0]).getStartArgNo(this.getClass()) : -1;
-
-			// 対象引数が-1の時は該当なしなのでコード補完しない
-			if (startArgNo == -1 || startArgNo >= len) {
-				return;
-			} else {
-				boolean isBlank = buffer.endsWith(" ");
-				if (len >= startArgNo) {
-					// sqlNameが指定されている場合
-					String sqlName = parts[startArgNo - 1];
-					if (StringUtils.isNotEmpty(sqlName)) {
-						String sql = config.getSqlManager().getSql(sqlName);
-						if (StringUtils.isNotEmpty(sql)) {
-							Set<String> params = getSqlParams(sql);
-							if (len > startArgNo) {
-								// 最後のパラメータ以外ですでに指定されたバインドパラメータを候補から除去する
-								int lastPos = isBlank ? len : len - 1;
-								for (int i = startArgNo; i < lastPos; i++) {
-									String part = parts[i];
-									String[] keyValue = part.split("=", 2);
-									params.remove(keyValue[0]);
-								}
-								if (isBlank) {
-									// 候補の表示位置を計算
-									candidates.addAll(params.stream().map(Candidate::new).collect(Collectors.toList()));
-								} else {
-									// 候補の表示位置を計算
-									pos = pos - parts[len - 1].length();
-									// 最後のパラメータについて候補を作成
-									String[] keyValue = parts[len - 1].split("=", 2);
-									if (keyValue.length == 2) {
-										// すでに値の入力があるため補完は行わない
-										pos = -1;
-									} else {
-										String key = keyValue[0];
-										for (String match : params) {
-											if (match.startsWith(key)) {
-												candidates.add(new Candidate(match));
-											}
-										}
-									}
-								}
-							} else {
-								candidates.addAll(params.stream().map(Candidate::new).collect(Collectors.toList()));
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	/**
-	 * テーブル名を補完するCompleter
-	 *
-	 * @author H.Sugimoto
-	 *
-	 */
-	class TableNameCompleter implements Completer {
-
-		/**
-		 * DatabaseMetadateからテーブル名を取得して補完候補とする。
-		 *
-		 * {@inheritDoc}
-		 *
-		 * @see org.jline.reader.Completer#complete(org.jline.reader.LineReader, org.jline.reader.ParsedLine, java.util.List)
-		 */
-		@Override
-		public void complete(final LineReader reader, final ParsedLine line, final List<Candidate> candidates) {
-			String buffer = line.line();
-			String[] parts = getLineParts(buffer);
-			int len = parts.length;
-
-			// コード補完する引数の番号を特定。
-			int startArgNo = len >= 1 ? Command.toCommand(parts[0]).getStartArgNo(this.getClass()) : -1;
-
-			// 対象引数が-1の時は該当なしなのでコード補完しない
-			if (startArgNo == -1) {
-				return;
-			} else {
-				boolean isBlank = buffer.endsWith(" ");
-				String tableNamePattern = "%";
-				if (len == startArgNo && isBlank) {
-					tableNamePattern = "%";
-				} else if (len == startArgNo + 1 && !isBlank) {
-					tableNamePattern = parts[len - 1] + "%";
-				} else {
-					return;
-				}
-
-				try {
-					Connection conn = config.getConnectionSupplier().getConnection();
-					DatabaseMetaData md = conn.getMetaData();
-					try (ResultSet rs = md.getTables(conn.getCatalog(), conn.getSchema(),
-							tableNamePattern.toUpperCase(), null)) {
-						while (rs.next()) {
-							candidates.add(new Candidate(rs.getString("TABLE_NAME")));
-						}
-					}
-					if (candidates.isEmpty()) {
-						try (ResultSet rs = md.getTables(conn.getCatalog(), conn.getSchema(),
-								tableNamePattern.toLowerCase(), null)) {
-							while (rs.next()) {
-								candidates.add(new Candidate(rs.getString("TABLE_NAME")));
-							}
-						}
-					}
-				} catch (SQLException ex) {
-					ex.printStackTrace();
-					return;
-				}
-			}
-		}
-	}
-
-	class SqlKeywordCompleter implements Completer {
-		@Override
-		public void complete(final LineReader reader, final ParsedLine line, final List<Candidate> candidates) {
-			String buffer = line.line();
-			String[] parts = getLineParts(buffer);
-			int len = parts.length;
-
-			// コード補完する引数の番号を特定。
-			int startArgNo = len >= 1 ? Command.toCommand(parts[0]).getStartArgNo(this.getClass()) : -1;
-
-			// 対象引数が-1の時は該当なしなのでコード補完しない
-			if (startArgNo == -1) {
-				return;
-			} else {
-				String key = null;
-
-				boolean isBlank = buffer.endsWith(" ");
-				if (len == startArgNo && isBlank) {
-					key = null;
-				} else if (len == startArgNo + 1 && !isBlank) {
-					key = parts[len - 1];
-				} else {
-					return;
-				}
-
-				final String keyword = key;
-
-				Stream.of("select", "insert", "update", "delete")
-						.filter(c -> keyword == null || c.startsWith(keyword.toLowerCase()))
-						.forEach(c -> candidates.add(new Candidate(c)));
-			}
-		}
-	}
-
-	/**
-	 * 動的ロードしたDriverの中継用クラス
-	 *
-	 * DriverManagerはclasspathで指定したクラスしか実行しないため、中継用のクラス経由で動的にロードしたDriverを実行する
-	 *
-	 * @author H.Sugimoto
-	 *
-	 */
-	class DriverShim implements Driver {
-		private Driver driver = null;
-
-		/**
-		 * コンストラクタ
-		 *
-		 * @param driver 実Driver
-		 */
-		DriverShim(final Driver driver) {
-			this.driver = driver;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see java.sql.Driver#connect(java.lang.String, java.util.Properties)
-		 */
-		@Override
-		public Connection connect(final String url, final Properties info) throws SQLException {
-			return driver.connect(url, info);
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see java.sql.Driver#acceptsURL(java.lang.String)
-		 */
-		@Override
-		public boolean acceptsURL(final String url) throws SQLException {
-			return driver.acceptsURL(url);
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see java.sql.Driver#getPropertyInfo(java.lang.String, java.util.Properties)
-		 */
-		@Override
-		public DriverPropertyInfo[] getPropertyInfo(final String url, final Properties info) throws SQLException {
-			return driver.getPropertyInfo(url, info);
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see java.sql.Driver#getMajorVersion()
-		 */
-		@Override
-		public int getMajorVersion() {
-			return driver.getMajorVersion();
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see java.sql.Driver#getMinorVersion()
-		 */
-		@Override
-		public int getMinorVersion() {
-			return driver.getMinorVersion();
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see java.sql.Driver#jdbcCompliant()
-		 */
-		@Override
-		public boolean jdbcCompliant() {
-			return driver.jdbcCompliant();
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see java.sql.Driver#getParentLogger()
-		 */
-		@Override
-		public java.util.logging.Logger getParentLogger() throws SQLFeatureNotSupportedException {
-			return driver.getParentLogger();
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see java.lang.Object#toString()
-		 */
-		@Override
-		public String toString() {
-			return driver.toString();
-		}
 	}
 
 	/**
